Movement Emitter for player
There's also prototypes for Landing Emitters and interactable flowers,
but the Landing Emitter will need help since the code currently doesn't
account for it and I really don't want to break the collision checking
code and the flower has this problem where it consistently messes with
player collision. I spent quite a bit of time today trying to get it
working today, maybe we can talk implementation and alternatives at the
next meeting.